fixed package paths
